flaskcli3000/base/api/flask/rest/business/sample_api/requirements.txt
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.2.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
Flask-Cors==3.0.8
flask-restplus==0.13.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pep8==1.7.1
pluggy==0.13.0
py==1.8.0
py-healthcheck==1.9.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.3
pytest-env==0.6.2
pytest-flask==0.15.0
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.12.0
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0